# docs/api/openapi.yaml
{
  "openapi": "3.1.0",
  "info": {
    "title": "Marketing Analytics",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/import/historical": {
      "post": {
        "tags": [
          "imports"
        ],
        "summary": "Import Historical",
        "operationId": "import_historical_api_import_historical_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HistoricalImportRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HistoricalImportResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/import/historical-csv": {
      "post": {
        "tags": [
          "imports"
        ],
        "summary": "Import Historical Csv",
        "operationId": "import_historical_csv_api_import_historical_csv_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HistoricalCsvImportRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HistoricalImportResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/import/historical-csv/preview": {
      "post": {
        "tags": [
          "imports"
        ],
        "summary": "Preview Historical Csv",
        "operationId": "preview_historical_csv_api_import_historical_csv_preview_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HistoricalCsvImportRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HistoricalImportPreviewResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/import/history": {
      "get": {
        "tags": [
          "imports"
        ],
        "summary": "Import History",
        "operationId": "import_history_api_import_history_get",
        "parameters": [
          {
            "name": "site_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Site Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HistoricalImportHistoryResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/import/batches/{batch_id}/rollback": {
      "post": {
        "tags": [
          "imports"
        ],
        "summary": "Rollback Import Batch",
        "operationId": "rollback_import_batch_api_import_batches__batch_id__rollback_post",
        "parameters": [
          {
            "name": "batch_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Batch Id"
            }
          },
          {
            "name": "site_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Site Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HistoricalImportRollbackResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "HistoricalCsvImportRequest": {
        "properties": {
          "site_id": {
            "type": "string",
            "title": "Site Id"
          },
          "csv_text": {
            "type": "string",
            "title": "Csv Text"
          }
        },
        "type": "object",
        "required": [
          "site_id",
          "csv_text"
        ],
        "title": "HistoricalCsvImportRequest"
      },
...
      "HistoricalImportHistoryResponse": {
        "properties": {
          "site_id": {
            "type": "string",
            "title": "Site Id"
          },
          "batches": {
            "items": {
              "$ref": "#/components/schemas/HistoricalImportBatchResponse"
            },
            "type": "array",
            "title": "Batches"
          }
        },
        "type": "object",
        "required": [
          "site_id",
          "batches"
        ],
        "title": "HistoricalImportHistoryResponse"
      },
      "HistoricalImportPreviewOverlap": {
        "properties": {
          "day": {
            "type": "string",
            "format": "date",
            "title": "Day"
          },
          "metric": {
            "type": "string",
            "title": "Metric"
          },
          "source": {
            "type": "string",
            "enum": [
              "live",
              "historical_import"
            ],
            "title": "Source"
          },
          "count": {
            "type": "integer",
            "title": "Count"
          }
        },
        "type": "object",
        "required": [
          "day",
          "metric",
          "source",
          "count"
        ],
        "title": "HistoricalImportPreviewOverlap"
      },
      "HistoricalImportPreviewResponse": {
        "properties": {
          "site_id": {
            "type": "string",
            "title": "Site Id"
          },
          "valid": {
            "type": "boolean",
            "title": "Valid"
          },
          "row_count": {
            "type": "integer",
            "title": "Row Count"
          },
          "day_count": {
            "type": "integer",
            "title": "Day Count"
          },
          "start_day": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Start Day"
          },
          "end_day": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "End Day"
          },
          "metrics": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Metrics"
          },
          "errors": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Errors"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Warnings"
          },
          "live_overlaps": {
            "items": {
              "$ref": "#/components/schemas/HistoricalImportPreviewOverlap"
            },
            "type": "array",
            "title": "Live Overlaps"
          },
          "replaceable_import_overlaps": {
            "items": {
              "$ref": "#/components/schemas/HistoricalImportPreviewOverlap"
            },
            "type": "array",
            "title": "Replaceable Import Overlaps"
          }
        },
        "type": "object",
        "required": [
          "site_id",
          "valid",
          "row_count",
          "day_count"
        ],
        "title": "HistoricalImportPreviewResponse"
      },
      "HistoricalImportRequest": {
        "properties": {
          "site_id": {
            "type": "string",
            "title": "Site Id"
          },
          "rows": {
            "items": {
              "$ref": "#/components/schemas/HistoricalImportRow"
            },
            "type": "array",
            "title": "Rows"
          }
        },
        "type": "object",
        "required": [
          "site_id",
          "rows"
        ],
        "title": "HistoricalImportRequest"
      },
      "HistoricalImportResponse": {
        "properties": {
          "site_id": {
            "type": "string",
            "title": "Site Id"
          },
          "imported_rows": {
            "type": "integer",
            "title": "Imported Rows"
          },
          "reduced_days": {
            "type": "integer",
            "title": "Reduced Days"
          },
          "batch_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Batch Id"
          }
        },
        "type": "object",
        "required": [
          "site_id",
          "imported_rows",
          "reduced_days"
        ],
        "title": "HistoricalImportResponse"
      },
      "HistoricalImportRollbackResponse": {
        "properties": {
          "site_id": {
            "type": "string",
            "title": "Site Id"
          },
          "batch_id": {
            "type": "integer",
            "title": "Batch Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "deleted_rows": {
            "type": "integer",
            "title": "Deleted Rows"
          },
          "reduced_days": {
            "type": "integer",
            "title": "Reduced Days"
          }
        },
        "type": "object",
        "required": [
          "site_id",
          "batch_id",
          "status",
          "deleted_rows",
          "reduced_days"
        ],
        "title": "HistoricalImportRollbackResponse"
      },
      "HistoricalImportRow": {
        "properties": {
          "day": {
            "type": "string",
            "format": "date",
            "title": "Day"
          },
          "metric": {
            "type": "string",
            "enum": [
              "uniques",
              "pageviews",
              "sessions",
              "conversions",
              "revenue"
            ],
            "title": "Metric"
          },
          "value": {
            "type": "number",
            "minimum": 0.0,
            "title": "Value"
          }
        },
        "type": "object",
        "required": [
          "day",
          "metric",
          "value"
        ],
        "title": "HistoricalImportRow"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}